+++ app/src/main/java/org/piramalswasthya/sakhi/model/dynamicEntity/FormSubmitRequest.kt
@@ -2,8 +2,8 @@ package org.piramalswasthya.sakhi.model.dynamicEntity
 
 data class FormSubmitRequest(
     val formId: String,
-    val beneficiaryId: Int,
-    val houseHoldId: Int,
+    val beneficiaryId: Long,
+    val houseHoldId: Long,
     val visitDate: String,
     val fields: Map<String, Any?>
 )
